Added MS Rewards Calculator --calculator
Created a calculator which allows you to calculate how long it will take to earn a certain amount of money using --calculator. Adjust options such as: number of accounts, current balance, and is if it is a Microsoft associated gift card or not. And shows excess balance in calculation result. modified implementation based on https://github.com/Prabh-Saini/msrewardsestimator.